move validation functions to utils
